Доработки UserDao.java
